% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_technical_replicates.R
\name{average_technical_replicates}
\alias{average_technical_replicates}
\title{Average technical replicates}
\usage{
average_technical_replicates(raw_data)
}
\arguments{
\item{raw_data}{A dataframe containing the raw fluorescence data. It must
contain at least 8 columns named
- \code{Experiment}: A unique string identifier for each experiment
- \code{Type}: Either "blank" (no donor) or "titration" (donor present)
- \code{Replicate}: An integer that indicates the technical replicate
number (ie, 1, 2,...)
- \code{Observation}: An integer that identifies each observation
within a run. For a plate-reader based assay, this is usually the row
number of a well. The number of observations for a titration and a
blank must match, and a given \code{Observation} value must have the
same ligand concentration for the titration and the blank.
- \code{concentration}: Ligand concentration
- \code{donor_channel}: Donor fluorescence
- \code{acceptor_channel}: Acceptor fluorescence
- \code{fret_channel}: FRET signal}
}
\value{
A dataframe containing the reduced dataset after averaging across
    replicates. It contains all of the above columns *except*
    \code{Replicate} (becuase we average over replicates)
}
\description{
This function calculates averages of fluorescence values from
    an arbitrary number of technical replicates.
}
