% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_process.R
\name{batch_process}
\alias{batch_process}
\title{Batch process FRET data}
\usage{
batch_process(directory_in = NULL, titrations = NULL, blanks = NULL,
  skip_lines = 4, directory_out = NULL)
}
\arguments{
\item{directory_in}{A directory containing raw data files.}

\item{titrations}{A character vector containing words that identify the
titration series (e.g. \code{titration}). These words must match the
ones in \code{raw_data$Content} (case sensitive).}

\item{blanks}{A character vector containing words that identify the blank
experiment series (e.g. \code{blank}). These words must match the ones
in \code{raw_data$Content} (case sensitive).}

\item{skip_lines}{Number of lines to skip at the beginning of CSV files (see
\code{\link[utils]{read.csv}}). Defaults to 4.}

\item{directory_out}{An optional directory where to write output files.}
}
\value{
Write to \code{directory_out} data files containing the corrected
    FRET signal from each input raw data file. Each output file contains two
    columns named \code{concentration} and \code{fret_corrected}, and has the
    same file name as the original raw data file appended with
    \code{_corrected}. If no output directory is specified, the function
    instead returns a list containing the corrected datasets, which can then
    be accessed programmatically.
}
\description{
This function applies \code{\link{average_technical_replicates}}
    and \code{\link{correct_fret_signal}} to all raw data files present in a
    given directory, and either writes output files in another directory or
    returns a list containing corrected datasets for programmatic
    manipulation.
}
\seealso{
\code{\link{average_technical_replicates}} and
    \code{\link{correct_fret_signal}}, which can be used manually on an
    individual dataset.
}
