% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_master_file.R
\name{cleanData}
\alias{cleanData}
\title{Clean data from single csv file}
\usage{
cleanData(fileName, numberOfLinesToSkip)
}
\arguments{
\item{fileName}{Name of single csv file to process.}

\item{numberOfLinesToSkip}{Number of lines to skip when reading the csv file, usually containing header information.}
}
\value{
One dataframe is returned, containing 11 columns with the input data from a single file: \code{Experiment},
\code{Type}, \code{Replicate}, \code{Observation}, \code{Well Row}, \code{Well Col}, \code{Content},
\code{fret_channel}, \code{donor_channel}, \code{acceptor_channel}, \code{concentration}.
}
\description{
This is an internal function that processes a single csv file and generates the
     \code{Experiment}, \code{Type}, \code{Replicate}, and \code{Observation} columns.  One data frame is
     returned with the rows of data generated from a single csv file.
}
\examples{
\dontrun{

 cleanData(file, 4)

 }

}
