% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_fret_signal.R
\name{correct_fret_signal}
\alias{correct_fret_signal}
\title{Correct FRET signal}
\usage{
correct_fret_signal(fret_data, F_donor_only, D_donor_only)
}
\arguments{
\item{fret_data}{A dataframe containing the fluorescence data (possibly an
average of technical replicates). It must contain at least four columns
named \code{fret_acceptor_only} (FRET channel of the blank titration with
only acceptor-labeled molecule), \code{acceptor_acceptor_only} (acceptor
channel of the blank titration with only acceptor-labeled molecule),
\code{fret_titration} (FRET channel of the actual titration) and
\code{concentration}.}

\item{F_donor_only}{A FRET channel fluorescence count measured from a sample
with donor only (typically, an average from 2 to 4 technical replicates).}

\item{D_donor_only}{A donor channel fluorescence count measured from a sample
with donor only (typically, an average from 2 to 4 technical replicates).}
}
\value{
A dataframe containing the corrected FRET signal. It contains two
  columns named \code{fret_corrected} and \code{concentration}.
}
\description{
This function subtracts the FRET background of the blank
  titration (without donor) from the raw FRET signal of the actual titration.
}
