% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_figure.R
\name{detect_binding_model}
\alias{detect_binding_model}
\title{Detect which binding model equation was used to fit a dataset}
\usage{
detect_binding_model(fit)
}
\arguments{
\item{fit}{An \code{\link[stats]{nls}} object. The output of
\code{\link{fit_binding_model}} can be used directly as input here.}
}
\value{
A character vector corresponding to the name of the binding model
    equation used to fit the dataset corresponding to the model object.
}
\description{
This internal function takes a model object from
    \code{\link{fit_binding_model}} or \code{\link[stats]{nls}} and returns
    the name of the binding model used to fit the dataset.
}
