% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/on_attach.R
\name{.onAttach}
\alias{.onAttach}
\title{Package startup messages}
\usage{
.onAttach(libname, pkgname)
}
\arguments{
\item{libname}{A character string giving the library directory where the
package defining the namespace was found. See \code{\link[base]{ns-hooks}}.}

\item{pkgname}{A character string giving the name of the package. See
\code{\link[base]{ns-hooks}}.}
}
\value{
Nothing, simply used for the side effect of displaying messages.
}
\description{
This internal function is run when the package gets attached by
    \code{\link[base]{library}}.
}
