% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_binding_model.R
\name{fit_hill}
\alias{fit_hill}
\title{Fit the Hill model equation to the experimental binding data}
\usage{
fit_hill(data)
}
\arguments{
\item{data}{A dataframe containing the binding signal. It must contain
at least two columns: \code{concentration} (the ligand concentration) and
\code{signal} (the observed binding signal).}
}
\value{
An \code{\link[stats]{nls}} object containing the results of the fit.
}
\description{
This internal function fits a the Hill model equation to the
    experimental binding data.
}
