% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_binding_model.R
\name{fit_quadratic}
\alias{fit_quadratic}
\title{Fit the quadratic model equation to the experimental binding data}
\usage{
fit_quadratic(data, probe_concentration)
}
\arguments{
\item{data}{A dataframe containing the binding signal. It must contain
at least two columns: \code{concentration} (the ligand concentration) and
\code{signal} (the observed binding signal).}

\item{probe_concentration}{The fixed concentration of probe molecule in the
experiment.}
}
\value{
An \code{\link[stats]{nls}} object containing the results of the fit.
}
\description{
This internal function fits a the quadratic model equation to
    the experimental binding data.
}
