% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fp_average_replicates.R
\name{fp_average_replicates}
\alias{fp_average_replicates}
\title{Average technical replicates from a fluorescence polarization or anisotropy experiment}
\usage{
fp_average_replicates(raw_data)
}
\arguments{
\item{raw_data}{A dataframe containing the fluorescence polarization or
anisotropy data. It must contain at least 8 columns named:
\describe{
\item{Experiment}{A unique name identifying each experiment.}
\item{Type}{For example, "titration".}
\item{Replicate}{A number identifying the technical replicate (1, 2, etc.).}
\item{Observation}{A number identifying each observation in a titration
series (corresponds to the plate column numbers, if experiments are set
up as rows in a 384-well plate). The number of observations for an
experiment and its blanks must match, and a given observation number must
associate data points at the same concentration in the titration series.}
\item{concentration}{The ligand concentration in the titration series.}
\item{polarization}{Fluorescence polarization.}
\item{anisotropy}{Fluorescence anisotropy.}
\item{intensity}{Fluorescence intensity.}
}
The output of \code{\link{fp_format_data}} can be used directly as input
for this function.}
}
\value{
A dataframe containing the reduced dataset after averaging across
    replicates. It contains all of the above columns \emph{except}
    \code{Replicate}, because it returns the average values over replicates.
}
\description{
This function calculates averages of fluorescence polarization
    or anisotropy values from an arbitrary number of technical replicates.
}
\seealso{
\code{\link{fp_format_data}} to prepare datasets for use with
    \code{fp_average_replicates}.
}
