% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fp_format_data.R
\name{fp_format_one_dataset}
\alias{fp_format_one_dataset}
\title{Format a single fluorescence polarization or anisotropy dataset for subsequent processing}
\usage{
fp_format_one_dataset(raw_data, experiment_name)
}
\arguments{
\item{raw_data}{A single dataframe to process.}

\item{experiment_name}{The name of the corresponding experiment.}
}
\value{
A dataframe containing 8 columns: \code{Experiment}, \code{Type},
    \code{Replicate}, \code{Observation}, \code{polarization},
    \code{anisotropy}, \code{intensity} and \code{concentration}.
}
\description{
This internal function processes a single dataframe to generate
    the \code{Experiment}, \code{Type}, \code{Replicate}, and
    \code{Observation} columns required for subsequent processing by
    \code{\link{fp_average_replicates}}.
}
\examples{
\dontrun{
fp_format_one_dataset(my_data, "my_experiment")
}

}
