% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fp_inspect_raw_data.R
\name{fp_save_inspection_plot}
\alias{fp_save_inspection_plot}
\title{Save raw data plots from a single FP binding dataset}
\usage{
fp_save_inspection_plot(input_plot, dataset_name, output_directory, plot_format)
}
\arguments{
\item{input_plot}{The output of  \code{\link{fp_inspect_one_dataset}}.}

\item{dataset_name}{The name of the corresponding dataset.}

\item{output_directory}{The name of the output directory where plots will be
saved.}

\item{plot_format}{A character string indicating the file format to use to
save plots. Possible values are \code{"png"} (default value),
\code{"pdf"} and \code{"svg"}.}
}
\value{
Writes plots in files on disk.
}
\description{
This internal function saves the total fluorescence intensity
    plots from an FP dataset to PNG, PDF or SVG files.
}
