% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fp_use_signal.R
\name{fp_use_signal}
\alias{fp_use_signal}
\title{Choose which signal to fit between fluorescence polarization and anisotropy}
\usage{
fp_use_signal(data, sig = NULL)
}
\arguments{
\item{data}{A dataframe containing the fluorescence polarization and
anisotropy data. It must contain at least 7 columns named:
\describe{
\item{Experiment}{A unique name identifying each experiment.}
\item{Type}{For example, "titration".}
\item{Observation}{A number identifying each observation in a titration
series (corresponds to the plate column numbers, if experiments are set
up as rows in a 384-well plate). The number of observations for an
experiment and its blanks must match, and a given observation number must
associate data points at the same concentration in the titration series.}
\item{concentration}{The ligand concentration in the titration series.}
\item{Polarization}{Fluorescence polarization.}
\item{Anisotropy}{Fluorescence anisotropy.}
\item{Intensity}{Fluorescence intensity.}
}
The output of \code{\link{fp_average_replicates}} can be used directly as
input for this function.}

\item{sig}{A character string to identify the signal to use for curve fitting.
Possible values are \code{"polarization"} or \code{"p"} to use fluorescence
polarization signal, and \code{"anisotropy"} or \code{"a"} to use
fluorescence anisotropy signal.}
}
\value{
A dataframe containing the same dataset with columns appropriately
    named for \code{\link{fit_binding_model}}.
}
\description{
This function renames the appropriate column in a fluorescence
    polarization and anisotropy dataset for subsequent fitting by
    \code{\link{fit_binding_model}}.
}
\seealso{
\code{\link{fp_format_data}} and \code{\link{fp_average_replicates}}.
}
