% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fret_correct_signal.R
\name{fret_correct_signal}
\alias{fret_correct_signal}
\title{Correct FRET signal}
\usage{
fret_correct_signal(data, output_directory = NULL)
}
\arguments{
\item{data}{A dataframe containing the FRET data after avergaing over
technical replicates. This dataframe must contain the following columns:
\describe{
\item{Experiment}{A unique name identifying each experiment.}
\item{Type}{Either "acceptor_only" (no donor control), "donor_only" (no
acceptor control), or "titration" (experiment with both donor and
acceptor).}
\item{Observation}{A number identifying each observation in a titration
series (corresponds to the plate column numbers, if experiments are set
up as rows in a 384-well plate). The number of observations for an
experiment and its blanks must match, and a given observation number must
associate data points at the same concentration in the titration series.}
\item{concentration}{The ligand concentration in the titration series.}
\item{fret_channel}{Fluorescence intensity in the FRET channel.}
\item{acceptor_channel}{Fluorescence intensity in the acceptor channel.}
\item{donor_channel}{Fluorescence intensity in the donor channel.}
}
The output of \code{\link{fret_average_replicates}} can be used
directly as input for this function.}

\item{output_directory}{An optional output directory name where to write
corrected data in CSV files. This directory will be created if it does
not already exist.}
}
\value{
A dataframe containing the corrected FRET signal. It contains three
    columns:
    \describe{
    \item{Experiment}{A unique name identifying each experiment (same as in
    the input data).}
    \item{concentration}{The ligand concentration in the titration series
    (same as in the input data).}
    \item{signal}{The corrected FRET signal.}
    }
}
\description{
This function corrects the raw FRET signal by applying
    corrections for donor bleedthrough and acceptor direct excitation. See
    \href{https://doi.org/10.1093/nar/gkr1045}{Hieb AR \emph{et al} (2012)}
    and \href{https://doi.org/10.1016/B978-0-12-391940-3.00011-1}{Winkler DD
    \emph{et al} (2012)} for details.
}
\seealso{
\code{\link{fret_average_replicates}} to prepare a dataset for
    use with \code{fret_correct_signal}.

    For details on the signal correction applied by \code{fret_correct_signal},
    see:
    \itemize{
    \item Hieb AR \emph{et al} (2012) Fluorescence Strategies for High-Throughput
    Quantification of Protein Interactions. \emph{Nucleic Acids Research}
    40 (5): e33 (\href{https://doi.org/10.1093/nar/gkr1045}{doi:10.1093/nar/gkr1045}) and
    \item Winkler DD \emph{et al} (2012) Quantifying Chromatin-Associated
    Interactions: The HI-FI System. In \emph{Methods in Enzymology} pp
    243–274. Elsevier
    (\href{https://doi.org/10.1016/B978-0-12-391940-3.00011-1}{doi:10.1016/B978-0-12-391940-3.00011-1}).
    }
}
