% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fret_inspect_raw_data.R
\name{fret_inspect_one_dataset}
\alias{fret_inspect_one_dataset}
\title{Inspect raw fluorescence data from a single FRET binding assay}
\usage{
fret_inspect_one_dataset(dataset, highest_signal = NULL)
}
\arguments{
\item{dataset}{A dataframe containing the raw fluorescence data. It must
be the output of \code{\link{format_data}}.}

\item{highest_signal}{An optional number corresponding to the maximal signal
measurable by the plate reader instrument used. Defaults to \code{NULL},
which won't check for the presence of saturated reads. The input number
is not checked in any way: make sure it really corresponds to a saturated
read with your instrument.}
}
\value{
A named list containing three \code{ggplot2} graph objects named
    \code{donor}, \code{acceptor} and \code{fret}.
}
\description{
This internal function plots the raw fluorescence data from the
    donor, acceptor and FRET channels, as a function of the titration series.
}
