% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fret_inspect_raw_data.R
\name{fret_inspect_raw_data}
\alias{fret_inspect_raw_data}
\title{Inspect raw fluorescence data from FRET binding assays}
\usage{
fret_inspect_raw_data(raw_data, highest_signal = NULL,
  output_directory = NULL, plot_format = "png")
}
\arguments{
\item{raw_data}{A dataframe containing the raw fluorescence data. It must
be the output of \code{\link{format_data}}.}

\item{highest_signal}{An optional number corresponding to the maximal signal
measurable by the plate reader instrument used. Defaults to \code{NULL},
which won't check for the presence of saturated reads. The input number
is not checked in any way: make sure it really corresponds to a saturated
read with your instrument.}

\item{output_directory}{An optional directory name where to write plots. This
directory will be created if it does not already exist.}

\item{plot_format}{A character string indicating the file format to use to
save plots. Possible values are \code{"png"} (default value),
\code{"pdf"} and \code{"svg"}.}
}
\value{
A list in which each item is named after the corresponding experiment,
    and holds a named list containing three \code{ggplot2} graph objects
    named \code{donor}, \code{acceptor} and \code{fret}.
}
\description{
This function plots the raw fluorescence data from the donor,
    acceptor and FRET channels, as a function of the titration series.
}
