% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fret_job_plot.R
\name{fret_job_plot}
\alias{fret_job_plot}
\title{Make a figure with FRET stoichiometry data (Job plot)}
\usage{
fret_job_plot(dataset, output_directory = NULL, plot_format = "png")
}
\arguments{
\item{dataset}{A dataframe containing at least three columns called
\code{Experiment}, \code{fret_channel} (or \code{signal}), and
\code{concentration}. This can be the output of
\code{\link{format_data}} or \code{\link{fret_correct_signal}}.}

\item{output_directory}{An optional directory name where to write the plot.
This directory will be created if it does not already exist.}

\item{plot_format}{A character string indicating the file format to use to
save plots. Possible values are \code{"png"} (default value),
\code{"pdf"} and \code{"svg"}.}
}
\value{
A \code{\link[ggplot2]{ggplot}} graph object of a Job plot figure.
}
\description{
This function takes raw or corrected FRET data from a
    FRET stoichiometry experiment (continuous variation method) and outputs a
    Job plot.
}
