% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{fret_saturated_reads}
\alias{fret_saturated_reads}
\title{A FRET binding assay dataset with saturated reads}
\format{A data frame with 96 rows and 7 columns. Rows contain two replicates
of two experiments each containing 24 data points. Columns are:
\describe{
  \item{Well.Row}{The row where the sample corresponding to this record was
  located into the 384-well microplate.}
  \item{Well.Col}{The column where the sample corresponding to this record
  was located into the 384-well microplate.}
  \item{Content}{A word describing the experiment and replicate this
  particular data point belongs to (\code{blank_1}, \code{blank_2},
  \code{titration_1}, \code{titration_2}).}
  \item{fret_channel}{The fluorescence intensities measured in the FRET
  channel.}
  \item{acceptor_channel}{The fluorescence intensities measured in the
  acceptor channel.}
  \item{donor_channel}{The fluorescence intensities measured in the donor
  channel.}
  \item{concentration}{The concentrations of acceptor-labeled protein in the
  titration series.}
}}
\source{
Guillaume Gaullier
}
\usage{
fret_saturated_reads
}
\description{
This dataset was collected on March 13th, 2017. It is a protein-protein
interaction assay, where one of the two proteins is labeled with Alexa 488
(donor fluorophore) and the other with Atto 647 (acceptor fluorophore). This
dataset has a couple saturated reads (fluorescence intensity of 260000), but
can still yield a binding curve and a binding constant of 124 +/- 18.
}
\keyword{datasets}
