% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_user_metadata.R
\name{get_user_metadata}
\alias{get_user_metadata}
\title{Load user-supplied metadata}
\usage{
get_user_metadata(user_json_file)
}
\arguments{
\item{user_json_file}{A JSON file describing mappings between internal names
and names in the actual data files. It can contain the following entries:
\describe{
\item{\code{content}}{The name of the column describing sample content.}
\item{\code{concentration}}{The name of the column containing the
concentration series.}
\item{\code{fret_channel}}{The name of the column containing fluorescence
intensities from the FRET channel.}
\item{\code{acceptor_channel}}{The name of the column containing
fluorescence intensities from the acceptor channel.}
\item{\code{donor_channel}}{The name of the column containing fluorescence
intensities from the donor channel.}
\item{\code{titration}}{The name describing points of the titration series.}
\item{\code{acceptor_only}}{The name describing points of the control
series with only acceptor.}
\item{\code{donor_only}}{The name describing points of the control series
with only donor.}
\item{\code{buffer_only}}{The name describing points of the control
series with only buffer.}
\item{\code{parallel}}{The name of the column containing fluorescence
intensities from the parallel polarization plane.}
\item{\code{perpendicular}}{The name of the column containing fluorescence
intensities from the perpendicular polarization plane.}
\item{\code{polarization}}{The name of the column containing fluorescence
polarization.}
\item{\code{anisotropy}}{The name of the column containing fluorescence
anisotropy.}
\item{\code{intensity}}{The name of the column containing fluorescence
intensity.}
\item{\code{baseline}}{The name describing points of the baseline series
in a fluorescence polarization or anisotropy dataset.}}}
}
\value{
An invisible copy of the original environment containing default
    metadata. As a side effect, updates the column and sample names in the
    package-wide environment.
}
\description{
This internal function loads user-supplied metadata from a JSON
    file.
}
