% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_parameters.R
\name{guess_parameters}
\alias{guess_parameters}
\title{Guess initial values for parameters of the binding model equations}
\usage{
guess_parameters(input_data)
}
\arguments{
\item{input_data}{A dataframe containing the corrected binding signal. It must
contain at least two columns: \code{concentration} (the ligand
concentration) and \code{signal} (the observed binding signal).}
}
\value{
A named list containing initial guesses for \code{kd},
    \code{signal_min} and \code{signal_max}.
}
\description{
This function provides reasonable initial guesses for \code{kd},
    \code{signal_min} and \code{signal_max} (parameters of the binding model
    equations used to fit the data).
}
