% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_quadratic_parameters.R
\name{guess_quadratic_parameters}
\alias{guess_quadratic_parameters}
\title{Guess initial values for parameters of the quadratic binding equation}
\usage{
guess_quadratic_parameters(fret_corr)
}
\arguments{
\item{fret_corr}{A dataframe containing the corrected FRET signal. It must
contain at least two columns named \code{fret_corrected} and
\code{concentration}.}
}
\value{
A list containing initial guesses for \code{kd}, \code{fret_min}
    and \code{fret_max}.
}
\description{
This function guesses initial values for \code{kd},
    \code{fret_min} and \code{fret_max} (the parameters of the quadratic
    binding equation), using the corrected FRET data.
}
