% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equations.R
\name{hill}
\alias{hill}
\title{Hill model equation}
\usage{
hill(conc, parameters)
}
\arguments{
\item{conc}{Ligand concentration.}

\item{parameters}{A named list of parameters to be refined against the
experimental data. Parameters for this equation are a binding constant
(\code{kd}), a Hill coefficient (\code{n}), and values of minimal and
maximal signal (\code{signal_min}, \code{signal_max}).}
}
\description{
This function sets up the Hill model equation for use by
    \code{\link{fit_binding_model}}.
}
