% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equations.R
\name{hyperbolic}
\alias{hyperbolic}
\title{Hyperbolic model equation}
\usage{
hyperbolic(conc, parameters)
}
\arguments{
\item{conc}{Ligand concentration.}

\item{parameters}{A named list of parameters to be refined against the
experimental data. Parameters for this equation are a binding constant
(\code{kd}), a Hill coefficient (\code{n}), and values of minimal and
maximal signal (\code{signal_min}, \code{signal_max}).}
}
\description{
This function sets up the hyperbolic model equation for use by
    \code{\link{fit_binding_model}}. If a Hill coefficient different from 1
    is specified in the set of parameters, this equation will describe a Hill
    model.
}
