% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_raw_data.R
\name{inspect_one_raw_data}
\alias{inspect_one_raw_data}
\title{Inspect raw fluorescence data from a single FRET binding assay}
\usage{
inspect_one_raw_data(dataset, highest_signal = NULL)
}
\arguments{
\item{dataset}{A dataframe containing the raw fluorescence data. It must
be the output of \code{\link{format_data}}.}

\item{highest_signal}{An optional number corresponding to the maximal signal
measurable by the plate reader instrument used. Defaults to \code{NULL},
which won't check for the presence of saturated reads. The input number
is not checked in any way: make sure it really corresponds to a saturated
read with your instrument.}
}
\value{
A named list containing: a logical value \code{saturated_reads}
    indicating the presence of saturated reads in the dataset (equals to
    \code{TRUE} if saturated reads are present, \code{FALSE} if there is no
    saturated read, or \code{NULL} if this was not tested), and three
    \code{ggplot2} graph objects (named \code{donor}, \code{acceptor} and
    \code{fret}).
}
\description{
This internal function plots the raw fluorescence data from the
    donor, acceptor and FRET channels, as a function of the titration series.
}
