% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_raw_data.R
\name{inspect_raw_data}
\alias{inspect_raw_data}
\title{Inspect raw data}
\usage{
inspect_raw_data(raw_data, data_type, highest_signal = NULL,
  output_directory = NULL, plot_format = "png")
}
\arguments{
\item{raw_data}{A dataframe containing the raw data. It must be the output of
\code{\link{format_data}} or \code{\link{fp_calculate_pola_aniso_int}}.}

\item{data_type}{A character string describing which type of data is supplied
for pre-processing. Currently supported options are \code{"fret"} and
\code{"fp"}.}

\item{highest_signal}{An optional number corresponding to the maximal signal
measurable by the instrument used. Defaults to \code{NULL}, which won't
check for the presence of saturated reads. The input number is not
checked in any way: make sure it really corresponds to a saturated
read with your instrument.}

\item{output_directory}{An optional directory name where to write plots. This
directory will be created if it does not already exist.}

\item{plot_format}{A character string indicating the file format to use to
save plots. Possible values are \code{"png"} (default value),
\code{"pdf"} and \code{"svg"}.}
}
\value{
A list in which each item is named after the corresponding experiment,
    and holds a single \code{ggplot2} graph objects or a list of several such
    objects.
}
\description{
This function plots the raw data of experiments for visual
    inspection.
}
