% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{load_data}
\alias{load_data}
\title{Check and load input data}
\usage{
load_data(input = NULL, skip_lines = NULL)
}
\arguments{
\item{input}{A directory name, a vector of file names, a single dataset
already loaded in memory as a dataframe, or a list of datasets already
loaded in memory as dataframes.}

\item{skip_lines}{The number of lines to skip at the beginning of CSV files
(usually containing header information before the actual data starts).}
}
\value{
A named list of dataframes to be further processed by
    \code{*_format_data} functions.
}
\description{
This internal function checks input data before subsequent
    processing by \code{*_format_data} functions. It will accept all data
    files in a given directory, or any number of data files specified by
    their names, or any number of dataframes supplied as a named list, or
    even a single dataframe supplied as is.
}
\examples{
\dontrun{
load_data("my_data_directory", 4)
}
}
