% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_figure.R
\name{make_one_figure}
\alias{make_one_figure}
\title{Make a figure with data and binding curve from a single experiment}
\usage{
make_one_figure(fit, experiment_name, binding_model, probe_conc = NULL)
}
\arguments{
\item{fit}{An \code{\link[stats]{nls}} object. The output of
\code{\link{fit_binding_model}} can be used directly as input here.}

\item{experiment_name}{The name of the corresponding experiment.}

\item{binding_model}{The binding model used to fit the data.}

\item{probe_conc}{The fixed concentration of probe molecule. This is ignored
for models obtained with the hyperbolic and hill equations, but required
for models obtained with the quadratic equation.}
}
\value{
A \code{\link[ggplot2]{ggplot}} graph object of a binding curve
    figure.
}
\description{
This internal function takes a model object from
    \code{\link{fit_binding_model}} or \code{\link[stats]{nls}} and returns
    a \code{\link[ggplot2]{ggplot}} graph object of a binding curve figure.
}
