% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_master_file.R
\name{makeMasterFile}
\alias{makeMasterFile}
\title{Create master data frame from input csv file(s)}
\usage{
makeMasterFile(listOfInputFiles, numberOfLinesToSkip)
}
\arguments{
\item{listOfInputFiles}{A list of input csv files to be processed.}

\item{numberOfLinesToSkip}{The number of lines to skip in the csv file, usually containing header information.}
}
\value{
One dataframe is returned, containing 11 columns with the combined input data: \code{Experiment},
\code{Type}, \code{Replicate}, \code{Observation}, \code{Well Row}, \code{Well Col}, \code{Content},
\code{fret_channel}, \code{donor_channel}, \code{acceptor_channel}, \code{concentration}.
}
\description{
This function creates one 'master' data frame by combining all the input csv files. The
     two input variables are 1) list of csv files to process and 2) number of lines in the csv file to skip.
     For each input file, this function creates the \code{Experiment} column based on the \code{file_name}
     for each file_name.csv provided as input.  The \code{Type} and \code{Replicate} are extrapolated
     from the \code{Content} column in the original data.  \code{Type} is either 'blank' or 'titration'.
     \code{Replicate} is the number of technical replicate.  For example, \code{Type} = 'blank' and
     \code{Replicate} = '1' if \code{Content} = "blank_1".  The data is then grouped by \code{Experiment},
     \code{Type}, and \code{Replicate} to determine the \code{Observation} number.  \code{Observation}
     number should match the \code{Well Col} number. \code{dplyr::full_join} is used to combine the rows
     from each input file.
}
\examples{
\dontrun{

makeMasterFile(files, 4)

}
}
