% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan_experiment.R
\name{plan_experiment}
\alias{plan_experiment}
\title{Plan a binding assay experiment}
\usage{
plan_experiment(kd, min_concentration = 0.1, max_concentration = 1000,
  binding_model = hyperbola, probe_conc = NULL, hill = NULL)
}
\arguments{
\item{kd}{A Kd value. This can also be a numerical vector of values, and in
this case the function will plot several binding curves on the same
graph: one for each Kd value.}

\item{min_concentration}{The minimal value of the concentration range to
simulate the binding curve over. Defaults to \code{1e-1}.}

\item{max_concentration}{The maximal value of the concentration range to
simulate the binding curve over. Defaults to \code{1000}.}

\item{binding_model}{A binding model equation. Possible values are
\code{hyperbola}, \code{quadratic} and \code{quadratic_homodimer}.
Defaults to \code{hyperbola}.}

\item{probe_conc}{Fixed concentration of probe molecule. If not specified,
the hyperbola binding model equation is used by default.}

\item{hill}{An optional Hill coefficient.}
}
\value{
A \code{ggplot2} graph object of the simulated binding curve.
}
\description{
This function takes a value of Kd, a minimal and maximal
    concentration of titrant a binding model equation, an optional probe
    concentration and an optional Hill coefficient, and outputs a plot of the
    theoretical resulting binding curve. This is useful for planning an
    experiment, when knowing the order of magnitude of Kd, to ensure that the
    experimental binding curve will display both the lower and upper plateaus.
}
