% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equations.R
\name{quadratic}
\alias{quadratic}
\title{Quadratic model equation}
\usage{
quadratic(conc, parameters)
}
\arguments{
\item{conc}{Ligand concentration.}

\item{parameters}{A named list of parameters to be refined against the
experimental data. Parameters for this equation are a binding constant
(\code{kd}), a probe concentration (\code{probe_conc}), and values of
minimal and maximal signal (\code{signal_min}, \code{signal_max}).}
}
\description{
This function sets up the quadratic model equation for use by
    \code{\link{fit_binding_model}}.
}
