% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equations.R
\name{quadratic_homodimer}
\alias{quadratic_homodimer}
\title{Quadratic homodimer equation}
\usage{
quadratic_homodimer(titrated_conc, parameters)
}
\arguments{
\item{titrated_conc}{Concentrations in the titration series.}

\item{parameters}{A named list of parameters to be refined against the
experimental data. Parameters for this equation are a binding constant
(\code{kd}), a probe concentration (\code{probe_conc}), and values of
minimal and maximal signal (\code{signal_min}, \code{signal_max}).}
}
\description{
This function sets up the quadratic homodimer equation for use
    by \code{\link{fit_binding_model}} or \code{\link[stats]{nls}}.
}
