% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfret-package.R
\docType{package}
\name{rfret}
\alias{rfret}
\alias{rfret-package}
\title{rfret: Analyze FRET binding data}
\description{
The rfret package provides functions to perform all steps of the analysis of
FRET binding data: inspection of raw data, averaging of technical replicates,
correction of the FRET signal using signal from a blank experiment, fitting
of a binding model equation to the experimental data to determine the binding
constant, making a final figure, and performing analysis in batch processing
mode.
}
