% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_corrected_fret.R
\name{save_corrected_fret}
\alias{save_corrected_fret}
\title{Save corrected FRET data}
\usage{
save_corrected_fret(input = NULL, output_directory = NULL)
}
\arguments{
\item{input}{A single dataframe containing corrected FRET data, like the
output of \code{\link{correct_fret_signal}}.}

\item{output_directory}{The name of an output directory. This directory will
be created if it does not already exist.}
}
\value{
Write to \code{output_directory} CSV data files containing the
    corrected FRET signal from each input experiment. Each output file
    contains three columns named \code{Experiment}, \code{concentration} and
    \code{fret_corrected}, and has the same file name as the original raw
    data file appended with \code{_corrected}.
}
\description{
This function saves the corrected FRET data produced by
    \code{\link{correct_fret_signal}} to CSV files in a specified output
    directory. It will keep original file names, and append \code{_corrected}
    after the file name and before the \code{.csv} extension.
}
\seealso{
\code{\link{format_data}}, \code{\link{average_technical_replicates}}
    and \code{\link{correct_fret_signal}}, which are the other steps of the
    FRET data processing pipeline.
}
