% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_raw_data.R
\name{save_inspection_plots}
\alias{save_inspection_plots}
\title{Save raw data plots from a single FRET binding dataset}
\usage{
save_inspection_plots(input_plots, dataset_name, output_directory, plot_format)
}
\arguments{
\item{input_plots}{The output of  \code{\link{inspect_one_raw_data}}.}

\item{dataset_name}{The name of the corresponding dataset.}

\item{output_directory}{The name of the output directory where plots will be
saved.}

\item{plot_format}{A character string indicating the file format to use to
save plots. Possible values are \code{"png"} (default value),
\code{"pdf"} and \code{"svg"}.}
}
\value{
Writes plots in files on disk.
}
\description{
This internal function saves the raw fluorescence plots from the
    donor, acceptor and FRET channels to PNG, PDF or SVG files.
}
