% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_technical_replicates.R
\name{average_technical_replicates}
\alias{average_technical_replicates}
\title{Average technical replicates}
\usage{
average_technical_replicates(raw_data, blanks, titrations)
}
\arguments{
\item{raw_data}{A dataframe containing the raw fluorescence data. It must
contain at least five columns named \code{Content}, \code{donor_channel},
\code{acceptor_channel}, \code{fret_channel} and \code{concentration}.
The \code{Content} column must contain, for each row, a word describing
which data series or replicate this row belongs to (like
\code{blank_1}, \code{titration_1}, etc.).}

\item{blanks}{A character vector containing words that identify the blank
experiment series (e.g. \code{blank}). These words must match the ones
in \code{raw_data$Content} (case sensitive).}

\item{titrations}{A character vector containing words that identify the
titration series (e.g. \code{titration}). These words must match the
ones in \code{raw_data$Content} (case sensitive).}
}
\value{
A dataframe containing the reduced dataset. It contains five columns
    named \code{Content}, \code{donor_channel}, \code{acceptor_channel},
    \code{fret_channel} and \code{concentration}. The \code{Content} column
    contains only two words: \code{"blank"} and \code{"titration"}.
}
\description{
This function calculates averages of fluorescence values from
    an arbitrary number of technical replicates.
}
