% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_binding_model.R
\name{fit_binding_model}
\alias{fit_binding_model}
\title{Fit a binding model equation to the experimental FRET data}
\usage{
fit_binding_model(corrected_data, binding_model = NULL, parameters,
  donor_concentration = 10)
}
\arguments{
\item{corrected_data}{A dataframe containing the corrected FRET signal. It
must contain at least two columns named \code{fret_corrected} and
\code{concentration}. The output of the \code{\link{correct_fret_signal}}
function can be used directly as input here.}

\item{binding_model}{A binding model equation to fit to the experimental
data. Possible values are \code{hyperbola}, \code{quadratic} and
\code{quadratic_homodimer}.}

\item{parameters}{A named list containing initial guesses for \code{kd},
\code{fret_min} and \code{fret_max}. The output of
\code{\link{guess_parameters}} can be used directly as input here.}

\item{donor_concentration}{The concentration of donor-labeled molecule.
Defaults to 10.}
}
\value{
An named list containing: \code{fit}, an \code{\link[stats]{nls}}
    object containing the results of the fit; \code{binding_model}, the
    equation used in the fitting procedure; and \code{donor_concentration},
    the donor concentration used in the fitting procedure.
    This list can be directly used as input to \code{\link{make_figure}}.
}
\description{
This function uses initial guesses of \code{kd}, \code{fret_min}
    and \code{fret_max} to fit a binding model equation to the corrected
    experimental FRET data. This function is a wrapper around
    \code{\link[stats]{nls}} and provides reasonnable default settings (i.e.
    refine only \code{kd}, \code{fret_min} and \code{fret_max}, and possibly
    the Hill coefficient \code{n}, but keep probe concentration fixed).
}
