% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_binding_model.R
\name{fit_binding_model}
\alias{fit_binding_model}
\title{Fit a binding model equation to the experimental binding data}
\usage{
fit_binding_model(data, binding_model = NULL, probe_concentration = NULL,
  output_directory = NULL)
}
\arguments{
\item{data}{A dataframe containing the binding signal. It must contain
at least two columns: \code{concentration} (the ligand concentration) and
\code{signal} (the observed binding signal). The output of the
\code{\link{fret_correct_signal}} function can be used directly as input
here.}

\item{binding_model}{A binding model equation to fit to the experimental
data. Possible values are \code{"hyperbolic"}, \code{"hill"} or
\code{"quadratic"}.}

\item{probe_concentration}{The fixed concentration of probe molecule
(required to use the quadratic binding model, ignored by other models).}

\item{output_directory}{An optional directory name where to write fit results.
This directory will be created if it does not already exist.}
}
\value{
A named list where each element is named after the corresponding
    experiment and holds an \code{\link[stats]{nls}} object containing the
    results of the fit. Only succesful fits will be returned. This list can
    be directly used as input for \code{\link{make_figure}}.
}
\description{
This function fits a binding model equation (hyperbolic model,
    Hill model, or quadratic model) to experimental binding data.
}
