% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_binding_model.R
\name{fit_binding_model}
\alias{fit_binding_model}
\title{Fit a binding model equation to the experimental FRET data}
\usage{
fit_binding_model(fret_corrected, donor_concentration = 10, parameters)
}
\arguments{
\item{fret_corrected}{A dataframe containing the corrected FRET signal. It
must contain at least two columns named \code{fret_corrected} and
\code{concentration}. The output of the \code{\link{correct_fret_signal}}
function can be used directly as input here.}

\item{donor_concentration}{The concentration of donor-labeled molecule.
Defaults to 10.}

\item{parameters}{A named list containing initial guesses for the parameters
of the equation associated to the binding model. The output of
\code{\link{guess_quadratic_parameters}} can be used directly as
input here.}
}
\value{
A list containing estimates and standard errors of the model
    parameters from \code{\link[stats]{nls}}, stored in a
    data frame (list item named \code{results}), and three \code{ggplot2}
    graph objects of the data points and fit curve (list item named
    \code{binding_curve}), residual plot (list item named
    \code{residual_plot}) and final figure with both plots (list item named
    \code{final_figure}).
}
\description{
This function uses initial guesses of a binding model equation
    parameters to fit this equation to the corrected experimental FRET data.
}
