% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{format_data}
\alias{format_data}
\title{Format experimental data for subsequent processing}
\usage{
format_data(input = NULL, skip_lines = 0)
}
\arguments{
\item{input}{A directory name, a vector of file names, or a list of datasets
already loaded in memory as dataframes.}

\item{skip_lines}{The number of lines to skip at the beginning of CSV files
(usually containing header information before the actual data starts).}
}
\value{
A single dataframe with the combined input data, containing 11
    columns: \code{Experiment}, \code{Type}, \code{Replicate},
    \code{Observation}, \code{Well Row}, \code{Well Col}, \code{Content},
    \code{fret_channel}, \code{acceptor_channel}, \code{donor_channel} and
    \code{concentration}.
}
\description{
This function formats data for subsequent processing by
    \code{\link{average_technical_replicates}}. It will accept all data files
    in a given directory, or any number of data files specified by their
    names, or any number of dataframes supplied as a named list.
}
\examples{
\dontrun{
format_data("my_data_directory", 4)
}
}
