% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fret_format_data.R
\name{fret_format_data}
\alias{fret_format_data}
\title{Format raw experimental FRET binding data for subsequent processing}
\usage{
fret_format_data(input = NULL, skip_lines = 0)
}
\arguments{
\item{input}{A directory name, a vector of file names, a single dataset
already loaded in memory as a dataframe, or a list of datasets already
loaded in memory as dataframes.}

\item{skip_lines}{The number of lines to skip at the beginning of CSV files
(usually containing header information before the actual data starts).}
}
\value{
A single dataframe with the combined input data, containing 8
    columns: \code{Experiment}, \code{Type}, \code{Replicate},
    \code{Observation}, \code{fret_channel}, \code{acceptor_channel},
    \code{donor_channel} and \code{concentration}. Rows with missing values
    will be dropped.
}
\description{
This function formats data for subsequent processing by
    \code{\link{fret_average_replicates}}. It will accept all data files
    in a given directory, or any number of data files specified by their
    names, or any number of dataframes supplied as a named list, or even a
    single dataframe supplied as is.
}
\examples{
\dontrun{
format_data("my_data_directory", 4)
}

}
