% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_parameters.R
\name{guess_parameters}
\alias{guess_parameters}
\title{Guess initial values for parameters of the binding model equations}
\usage{
guess_parameters(fret_corr)
}
\arguments{
\item{fret_corr}{A dataframe containing the corrected FRET signal. It must
contain at least two columns named \code{fret_corrected} and
\code{concentration}.}
}
\value{
A list containing initial guesses for \code{kd}, \code{signal_min}
    and \code{signal_max}.
}
\description{
This function guesses initial values for \code{kd},
    \code{signal_min} and \code{signal_max} (the parameters of the binding
    model equations), using the corrected FRET data.
}
