% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan_experiment.R
\name{plan_experiment}
\alias{plan_experiment}
\title{Plan a binding assay experiment}
\usage{
plan_experiment(kd, min_concentration = 0.001,
  max_concentration = 1e+05, binding_model = "hyperbolic",
  probe_conc = NULL, hill_coef = 1)
}
\arguments{
\item{kd}{A Kd value.}

\item{min_concentration}{The minimal value of the concentration range to
simulate the binding curve over. Defaults to \code{1e-3}.}

\item{max_concentration}{The maximal value of the concentration range to
simulate the binding curve over. Defaults to \code{1e5}.}

\item{binding_model}{A binding model equation. Possible values are
\code{"hyperbolic"}, \code{"hill"} and \code{"quadratic"}. Defaults to
\code{"hyperbolic"}.}

\item{probe_conc}{Fixed concentration of probe molecule. If not specified,
the hyperbolic binding model equation is used by default.}

\item{hill_coef}{An optional Hill coefficient. This is ignored by the
quadratic binding model equation. Defaults to 1.}
}
\value{
A \code{ggplot2} graph object of the simulated binding curve.
}
\description{
This function takes a value of Kd, a minimal and maximal
    concentration of titrant a binding model equation, an optional probe
    concentration and an optional Hill coefficient, and outputs a plot of the
    theoretical resulting binding curve. This is useful for planning an
    experiment, when knowing the order of magnitude of Kd, to ensure that the
    experimental binding curve will display both the lower and upper plateaus.
}
