% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equations.R
\name{quadratic}
\alias{quadratic}
\title{Quadratic equation}
\usage{
quadratic(titrated_conc, kd, signal_min, signal_max, probe_conc)
}
\arguments{
\item{titrated_conc}{Concentrations in the titration series.}

\item{kd}{Binding constant.}

\item{signal_min}{Minimal observed signal.}

\item{signal_max}{Maximal observed signal.}

\item{probe_conc}{Fixed concentration of probe.}
}
\description{
This function sets up the quadratic equation for use by
    \code{\link{fit_binding_model}} or \code{\link[stats]{nls}}.
}
