% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equations.R
\name{quadratic_homodimer}
\alias{quadratic_homodimer}
\title{Quadratic homodimer equation}
\usage{
quadratic_homodimer(titrated_conc, kd, n, signal_min, signal_max, probe_conc)
}
\arguments{
\item{titrated_conc}{Concentrations in the titration series.}

\item{kd}{Binding constant.}

\item{n}{Not used in this equation.}

\item{signal_min}{Minimal observed signal.}

\item{signal_max}{Maximal observed signal.}

\item{probe_conc}{Fixed concentration of probe.}
}
\description{
This function sets up the quadratic homodimer equation for use
    by \code{\link{fit_binding_model}} or \code{\link[stats]{nls}}.
}
