% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{read_files}
\alias{read_files}
\title{Read files}
\usage{
read_files(input = NULL, skip_lines = NULL)
}
\arguments{
\item{input}{A vector of file names.}

\item{skip_lines}{The number of lines to skip at the beginning of CSV files
(usually containing header information before the actual data starts).}
}
\value{
A named list of dataframes to be further processed by
    \code{*_format_data} functions.
}
\description{
This internal function simply reads in data files.
}
\examples{
\dontrun{
load_data("my_data_directory", 4)
}
}
